+++ PurAir/src/com/philips/cl/di/dev/pa/dashboard/OutdoorFragment.java
@@ -110,7 +110,7 @@ public class OutdoorFragment extends BaseFragment implements OnClickListener, On
 		Drawable drawable = aqiPointer.getDrawable();
 		ALog.i(ALog.DASHBOARD, "OutdoorFragment$getRotationAnimation rotation " + rotation + " aqiPointer.getWidth()/2 " + (aqiPointer.getWidth()/2) + " drawable " + drawable.getMinimumHeight());
 		
-		Animation aqiCircleRotateAnim = new RotateAnimation(0.0f, rotation, drawable.getMinimumWidth()/2, drawable.getMinimumHeight()/2);
+		Animation aqiCircleRotateAnim = new RotateAnimation(0.0f, rotation, drawable.getMinimumWidth() / (float) 2, drawable.getMinimumHeight() / (float) 2);
 		
 	    aqiCircleRotateAnim.setDuration(2000);  
 	    aqiCircleRotateAnim.setRepeatCount(0);
